2018-07-13 修改administer

+++ src/main/java/com/web/administer/entity/Administer.java
@@ -4,15 +4,15 @@ package com.web.administer.entity;
  * 管理员实体类
  */
 public class Administer {
-    private Long admiid;    //管理员ID
+    private Integer admiid;    //管理员ID
     private String name;    //管理员用户名
     private String psw;     //管理员密码
 
-    public Long getAdmiid() {
+    public Integer getAdmiid() {
         return admiid;
     }
 
-    public void setAdmiid(Long admiid) {
+    public void setAdmiid(Integer admiid) {
         this.admiid = admiid;
     }
